Updated entrypoints to enable sharding based on env variable.
