+++ src/entrypoint/startSharded.ts
@@ -1,11 +1,11 @@
 import path from 'path';
 import { ShardingManager } from 'discord.js';
 import logger from '@moonstar-x/logger';
-import { DISCORD_TOKEN } from '../config/app';
+import { DISCORD_SHARDING_COUNT, DISCORD_TOKEN } from '../config/app';
 import { runMigrations } from '../app/migration';
 
 // Note: This works only when built, not in TypeScript.
-const startScript = path.join(__dirname, `./start.js`);
+const startScript = path.join(__dirname, `./startSingle.js`);
 
 const manager = new ShardingManager(startScript, { token: DISCORD_TOKEN });
 
@@ -15,5 +15,5 @@ manager.on('shardCreate', (shard) => {
 
 runMigrations()
   .then(() => {
-    manager.spawn({ amount: 'auto' });
+    manager.spawn({ amount: DISCORD_SHARDING_COUNT });
   });
